api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
google-genai==1.68.0
requests==2.32.3
python-dotenv==1.0.1
pydantic[email]==2.10.4
